moved database setup function from database.js to dbSetup.js to cleanup database file. We may or may not want to use that function later so saving it in it's own file for now.
