+++ database/database.js
@@ -1,43 +1,7 @@
-
 import pkg from 'pg';
-const { Client, Pool } = pkg;
+const { Pool } = pkg;
 import * as dotenv from "dotenv";
 dotenv.config({});
-
-const DB_NAME = process.env.DB_NAME || 'db-name';
-const DB_USER = process.env.DB_USER || 'postgres';
-const DB_HOST = process.env.DB_HOST || 'localhost';
-const DB_PASSWORD = process.env.DB_PASSWORD || 'root';
-
-async function setupDatabase() {
-
-    const isDev = process.env.NODE_ENV === 'development';
-
-    if (!isDev) return console.log('in production environment - skipping database creation.');
-
-    const client = new Client({
-        host: DB_HOST,
-        user: DB_USER,
-        password: DB_PASSWORD,
-        port: 5432,
-    });
-    
-    await client.connect();
-    
-    const res = await client.query(`SELECT datname FROM pg_catalog.pg_database WHERE datname = '${DB_NAME}'`);
-    
-    if (res.rowCount === 0) {
-        console.log(`${DB_NAME} database not found, creating it.`);
-        await client.query(`CREATE DATABASE "${DB_NAME}";`);
-        console.log(`created database ${DB_NAME}.`);
-    } else {
-        console.log(`${DB_NAME} database already exists.`);
-    }
-    
-    await client.end();
-}
-
-setupDatabase();
 
 const credentials = {
     user: process.env.DB_USER,
